pyproject.toml
[project]
name = "fastapi-saml"
version = "1.0.0"
description = "Example FastAPI backend that implements SAML2 authentication"
authors = [
    { name = "santigandolfo", email = "santiago.v.gandolfo@gmail.com" }
]
dependencies = [
    "fastapi[all]>=0.109.2",
    "pydantic>=2.6.1",
    "uvicorn[standard]>=0.27.1",
    "pydantic-settings>=2.1.0",
    "pysaml2>=7.3.1",
    "python-multipart>=0.0.9",
    "python-jose>=3.3.0",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
fix = true
unsafe-fixes = true

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "N",      # pep8-naming
    "I",      # isort
    "UP",     # pyupgrade
    "C4",     # flake8-comprehensions
    "B",      # flake8-bugbear
    "Q",      # flake8-quotes
    "C90",    # flake8-quotes
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit
    "COM",    # flake8-commas
    "T10",    # flake8-debugger
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "INP",    # flake8-no-pep420
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "ARG",    # flake8-unused-arguments
    "TD",     # flake8-todos
    "ERA",    # eradicate
    "FLY002", # flynt
    "RUF",    # Ruff-specific rules
]

ignore = [
    "E501",
    "B008",
    "C901",
    "B017",
    "N818",
    "S106",
    "ANN101",
    "ANN102",
    "COM812",
    "ISC001"
]

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.format]
line-ending = "lf"
